IGNITE-6858 Wait for exchange inside GridReduceQueryExecutor.query which never finishes due to opened transaction.
