+++ modules/core/src/main/resources/META-INF/classnames.properties
@@ -2095,3 +2095,4 @@ org.apache.ignite.transactions.TransactionState
 org.apache.ignite.transactions.TransactionTimeoutException
 org.apache.ignite.util.AttributeNodeFilter
 org.apache.ignite.internal.processors.cache.persistence.file.AsyncFileIO
+org.apache.ignite.cache.query.QueryRestartRequiredException
